requirements.txt
streamlit
python-dotenv
pypdf
langchain
faiss-cpu
langchain-google-genai
langchain-community